refactor : Producer Job 분리
