+++ cs25-batch/src/main/java/com/example/cs25batch/batch/jobs/DailyMailSendJob.java
@@ -89,57 +89,13 @@ public class DailyMailSendJob {
                     //log.info("[3. 문제 출제] QuizId : {} {}ms", quiz.getId(), quizEnd - quizStart);
 
                     //long mailStart = System.currentTimeMillis();
-                    mailService.sendQuizEmail(subscription, quiz);
+                    //mailService.sendQuizEmail(subscription, quiz);
                     //long mailEnd = System.currentTimeMillis();
                     //log.info("[4. 메일 발송] {}ms", mailEnd - mailStart);
                 }
             }
 
             log.info("[배치 종료] 메일 발송 완료");
-            return RepeatStatus.FINISHED;
-        };
-    }
-
-    //Message Queue 적용 후
-    @Bean
-    public Job mailProducerJob(JobRepository jobRepository,
-        @Qualifier("mailProducerStep") Step mailStep) {
-        return new JobBuilder("mailProducerJob", jobRepository)
-            .incrementer(new RunIdIncrementer())
-            .start(mailStep)
-            .build();
-    }
-
-    @Bean
-    public Step mailProducerStep(JobRepository jobRepository,
-        @Qualifier("mailProducerTasklet") Tasklet mailTasklet,
-        PlatformTransactionManager transactionManager) {
-        return new StepBuilder("mailProducerStep", jobRepository)
-            .tasklet(mailTasklet, transactionManager)
-            .build();
-    }
-
-    // TODO: Chunk 방식 고려
-    @Bean
-    public Tasklet mailProducerTasklet() {
-        return (contribution, chunkContext) -> {
-            log.info("[배치 시작] 메일 발송 대상 구독자 선별");
-
-            //long searchStart = System.currentTimeMillis();
-            List<SubscriptionMailTargetDto> subscriptions = subscriptionService.getTodaySubscriptions();
-            //long searchEnd = System.currentTimeMillis();
-            //log.info("[1. 발송 리스트 조회] {}개, {}ms", subscriptions.size(), searchEnd - searchStart);
-
-            for (SubscriptionMailTargetDto sub : subscriptions) {
-                Long subscriptionId = sub.getSubscriptionId();
-                //메일을 발송해야 할 구독자 정보를 MessageQueue 에 넣음
-                //long queueStart = System.currentTimeMillis();
-                batchProducerService.enqueueQuizEmail(subscriptionId);
-                //long queueEnd = System.currentTimeMillis();
-                //log.info("[2. Queue에 넣기] {}ms", queueEnd-queueStart);
-            }
-
-            log.info("[배치 종료] MessageQueue push 완료");
             return RepeatStatus.FINISHED;
         };
     }
